Address circular dependencies across multiple files
