+++ apps/studio/components/grid/components/grid/RowRenderer.tsx
@@ -2,8 +2,8 @@ import type { Key } from 'react'
 import { TriggerEvent, useContextMenu } from 'react-contexify'
 import { RenderRowProps, Row } from 'react-data-grid'
 
+import { ROW_CONTEXT_MENU_ID } from 'components/grid/constants'
 import { SupaRow } from 'components/grid/types'
-import { ROW_CONTEXT_MENU_ID } from '../menu'
 
 export default function RowRenderer(key: Key, props: RenderRowProps<SupaRow>) {
   const { show: showContextMenu } = useContextMenu()
