+++ apps/studio/components/grid/components/menu/index.ts
@@ -1,4 +1,2 @@
 export { default as ColumnMenu } from './ColumnMenu'
 export { default as RowContextMenu } from './RowContextMenu'
-
-export const ROW_CONTEXT_MENU_ID = 'row-context-menu-id'
