+++ apps/studio/components/interfaces/Integrations/CronJobs/CreateCronJobSheet/CreateCronJobSheet.tsx
@@ -1,15 +1,12 @@
 import { zodResolver } from '@hookform/resolvers/zod'
 import { PermissionAction } from '@supabase/shared-types/out/constants'
-import { toString as CronToString } from 'cronstrue'
 import { parseAsString, useQueryState } from 'nuqs'
 import { useEffect, useState } from 'react'
 import { SubmitHandler, useForm } from 'react-hook-form'
 import { toast } from 'sonner'
-import z from 'zod'
 
 import { useWatch } from '@ui/components/shadcn/ui/form'
 import { useParams } from 'common'
-import { urlRegex } from 'components/interfaces/Auth/Auth.constants'
 import EnableExtensionModal from 'components/interfaces/Database/Extensions/EnableExtensionModal'
 import { ButtonTooltip } from 'components/ui/ButtonTooltip'
 import { getDatabaseCronJob } from 'data/database-cron-jobs/database-cron-job-query'
@@ -38,136 +35,28 @@ import {
 import { Admonition } from 'ui-patterns/admonition'
 import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
 import { FormItemLayout } from 'ui-patterns/form/FormItemLayout/FormItemLayout'
-import { CRONJOB_DEFINITIONS } from './CronJobs.constants'
+import { CRONJOB_DEFINITIONS } from '../CronJobs.constants'
+import { buildCronQuery, buildHttpRequestCommand, parseCronJobCommand } from '../CronJobs.utils'
+import { EdgeFunctionSection } from '../EdgeFunctionSection'
+import { HttpBodyFieldSection } from '../HttpBodyFieldSection'
+import { HTTPHeaderFieldsSection } from '../HttpHeaderFieldsSection'
+import { HttpRequestSection } from '../HttpRequestSection'
+import { SqlFunctionSection } from '../SqlFunctionSection'
+import { SqlSnippetSection } from '../SqlSnippetSection'
 import {
-  buildCronQuery,
-  buildHttpRequestCommand,
-  cronPattern,
-  parseCronJobCommand,
-  secondsPattern,
-} from './CronJobs.utils'
+  FormSchema,
+  type CreateCronJobForm,
+  type CronJobType,
+} from './CreateCronJobSheet.constants'
 import { CronJobScheduleSection } from './CronJobScheduleSection'
-import { EdgeFunctionSection } from './EdgeFunctionSection'
-import { HttpBodyFieldSection } from './HttpBodyFieldSection'
-import { HTTPHeaderFieldsSection } from './HttpHeaderFieldsSection'
-import { HttpRequestSection } from './HttpRequestSection'
-import { SqlFunctionSection } from './SqlFunctionSection'
-import { SqlSnippetSection } from './SqlSnippetSection'
 
-export interface CreateCronJobSheetProps {
+interface CreateCronJobSheetProps {
   selectedCronJob?: Pick<CronJob, 'jobname' | 'schedule' | 'active' | 'command'>
   supportsSeconds: boolean
   isClosing: boolean
   setIsClosing: (v: boolean) => void
   onClose: () => void
 }
-
-const edgeFunctionSchema = z.object({
-  type: z.literal('edge_function'),
-  method: z.enum(['GET', 'POST']),
-  edgeFunctionName: z.string().trim().min(1, 'Please select one of the listed Edge Functions'),
-  timeoutMs: z.coerce.number().int().gte(1000).lte(5000).default(1000),
-  httpHeaders: z.array(z.object({ name: z.string(), value: z.string() })),
-  httpBody: z
-    .string()
-    .trim()
-    .optional()
-    .refine((value) => {
-      if (!value) return true
-      try {
-        JSON.parse(value)
-        return true
-      } catch {
-        return false
-      }
-    }, 'Input must be valid JSON'),
-  // When editing a cron job, we want to keep the original command as a snippet in case the user wants to manually edit it
-  snippet: z.string().trim(),
-})
-
-const httpRequestSchema = z.object({
-  type: z.literal('http_request'),
-  method: z.enum(['GET', 'POST']),
-  endpoint: z
-    .string()
-    .trim()
-    .min(1, 'Please provide a URL')
-    .regex(urlRegex(), 'Please provide a valid URL')
-    .refine((value) => value.startsWith('http'), 'Please include HTTP/HTTPs to your URL'),
-  timeoutMs: z.coerce.number().int().gte(1000).lte(5000).default(1000),
-  httpHeaders: z.array(z.object({ name: z.string(), value: z.string() })),
-  httpBody: z
-    .string()
-    .trim()
-    .optional()
-    .refine((value) => {
-      if (!value) return true
-      try {
-        JSON.parse(value)
-        return true
-      } catch {
-        return false
-      }
-    }, 'Input must be valid JSON'),
-  // When editing a cron job, we want to keep the original command as a snippet in case the user wants to manually edit it
-  snippet: z.string().trim(),
-})
-
-const sqlFunctionSchema = z.object({
-  type: z.literal('sql_function'),
-  schema: z.string().trim().min(1, 'Please select one of the listed database schemas'),
-  functionName: z.string().trim().min(1, 'Please select one of the listed database functions'),
-  // When editing a cron job, we want to keep the original command as a snippet in case the user wants to manually edit it
-  snippet: z.string().trim(),
-})
-
-const sqlSnippetSchema = z.object({
-  type: z.literal('sql_snippet'),
-  snippet: z.string().trim().min(1),
-})
-
-const FormSchema = z
-  .object({
-    name: z.string().trim().min(1, 'Please provide a name for your cron job'),
-    supportsSeconds: z.boolean(),
-    schedule: z
-      .string()
-      .trim()
-      .min(1)
-      .refine((value) => {
-        if (cronPattern.test(value)) {
-          try {
-            CronToString(value)
-            return true
-          } catch {
-            return false
-          }
-        } else if (secondsPattern.test(value)) {
-          return true
-        }
-        return false
-      }, 'Invalid Cron format'),
-    values: z.discriminatedUnion('type', [
-      edgeFunctionSchema,
-      httpRequestSchema,
-      sqlFunctionSchema,
-      sqlSnippetSchema,
-    ]),
-  })
-  .superRefine((data, ctx) => {
-    if (!cronPattern.test(data.schedule)) {
-      if (!(data.supportsSeconds && secondsPattern.test(data.schedule))) {
-        ctx.addIssue({
-          code: z.ZodIssueCode.custom,
-          message: 'Seconds are supported only in pg_cron v1.5.0+. Please use a valid Cron format.',
-          path: ['schedule'],
-        })
-      }
-    }
-  })
-
-export type CreateCronJobForm = z.infer<typeof FormSchema>
-export type CronJobType = CreateCronJobForm['values']
 
 const FORM_ID = 'create-cron-job-sidepanel'
