+++ apps/studio/components/interfaces/Integrations/CronJobs/CronJobPage.tsx
@@ -20,7 +20,7 @@ import {
   TooltipTrigger,
 } from 'ui'
 import ShimmeringLoader from 'ui-patterns/ShimmeringLoader'
-import { CreateCronJobSheet } from './CreateCronJobSheet'
+import { CreateCronJobSheet } from './CreateCronJobSheet/CreateCronJobSheet'
 import { isSecondsFormat, parseCronJobCommand } from './CronJobs.utils'
 import { PreviousRunsTab } from './PreviousRunsTab'
