+++ apps/studio/components/interfaces/Integrations/CronJobs/EdgeFunctionSection.tsx
@@ -21,7 +21,7 @@ import {
   SheetSection,
 } from 'ui'
 import { FormItemLayout } from 'ui-patterns/form/FormItemLayout/FormItemLayout'
-import { CreateCronJobForm } from './CreateCronJobSheet'
+import { CreateCronJobForm } from './CreateCronJobSheet/CreateCronJobSheet.constants'
 
 interface HTTPRequestFieldsProps {
   form: UseFormReturn<CreateCronJobForm>
