+++ apps/studio/components/interfaces/Integrations/CronJobs/HttpRequestSection.tsx
@@ -15,7 +15,7 @@ import {
   SheetSection,
 } from 'ui'
 import { FormItemLayout } from 'ui-patterns/form/FormItemLayout/FormItemLayout'
-import { CreateCronJobForm } from './CreateCronJobSheet'
+import { CreateCronJobForm } from './CreateCronJobSheet/CreateCronJobSheet.constants'
 
 interface HttpRequestSectionProps {
   form: UseFormReturn<CreateCronJobForm>
