+++ apps/studio/components/interfaces/Integrations/CronJobs/SqlFunctionSection.tsx
@@ -4,7 +4,7 @@ import FunctionSelector from 'components/ui/FunctionSelector'
 import SchemaSelector from 'components/ui/SchemaSelector'
 import { FormField_Shadcn_, SheetSection } from 'ui'
 import { FormItemLayout } from 'ui-patterns/form/FormItemLayout/FormItemLayout'
-import { CreateCronJobForm } from './CreateCronJobSheet'
+import { CreateCronJobForm } from './CreateCronJobSheet/CreateCronJobSheet.constants'
 
 interface SqlFunctionSectionProps {
   form: UseFormReturn<CreateCronJobForm>
