+++ apps/studio/components/interfaces/Integrations/CronJobs/SqlSnippetSection.tsx
@@ -3,7 +3,7 @@ import { UseFormReturn } from 'react-hook-form'
 import CodeEditor from 'components/ui/CodeEditor/CodeEditor'
 import { FormField_Shadcn_, SheetSection } from 'ui'
 import { FormItemLayout } from 'ui-patterns/form/FormItemLayout/FormItemLayout'
-import { CreateCronJobForm } from './CreateCronJobSheet'
+import { CreateCronJobForm } from './CreateCronJobSheet/CreateCronJobSheet.constants'
 
 interface SqlSnippetSectionProps {
   form: UseFormReturn<CreateCronJobForm>
