+++ apps/studio/components/interfaces/Integrations/Vault/Secrets/Secrets.types.ts
@@ -1,0 +1,7 @@
+export interface SecretTableColumn {
+  id: 'secret' | 'id' | 'secret_value' | 'updated_at' | 'actions'
+  name: string
+  minWidth?: number
+  width?: number
+  maxWidth?: number
+}
