+++ apps/studio/components/interfaces/Integrations/Wrappers/OverviewTab.tsx
@@ -18,6 +18,7 @@ import {
   WarningIcon,
 } from 'ui'
 import { IntegrationOverviewTab } from '../Integration/IntegrationOverviewTab'
+import { CreateIcebergWrapperSheet } from './CreateIcebergWrapperSheet'
 import { CreateWrapperSheet } from './CreateWrapperSheet'
 import { WRAPPERS } from './Wrappers.constants'
 import { WrapperTable } from './WrapperTable'
@@ -53,7 +54,13 @@ export const WrapperOverviewTab = () => {
   const databaseNeedsUpgrading =
     wrappersExtension?.installed_version === wrappersExtension?.default_version
 
-  const CreateWrapperSheetComponent = wrapperMeta.createComponent || CreateWrapperSheet
+  // [Joshen] Opting to declare custom wrapper sheets here instead of within Wrappers.constants.ts
+  // as we'll easily run into circular dependencies doing so unfortunately
+  const CreateWrapperSheetComponent = wrapperMeta.customComponent
+    ? wrapperMeta.name === 'iceberg_wrapper'
+      ? CreateIcebergWrapperSheet
+      : ({}) => null
+    : CreateWrapperSheet
 
   return (
     <IntegrationOverviewTab
