+++ apps/studio/components/interfaces/Integrations/Wrappers/Wrappers.types.ts
@@ -1,5 +1,3 @@
-import { CreateWrapperSheetProps } from './CreateWrapperSheet'
-
 export type WrapperMeta = {
   name: string
   handlerName: string
@@ -12,7 +10,7 @@ export type WrapperMeta = {
   server: Server
   docsUrl: string
   minimumExtensionVersion?: string
-  createComponent?: React.ComponentType<CreateWrapperSheetProps>
+  customComponent?: boolean
   // If true, the wrapper can target a schema which will be populated with tables specified by the wrapper..
   canTargetSchema?: boolean
   sourceSchemaOption?: ServerOption
