+++ apps/studio/components/interfaces/Reports/Reports.types.ts
@@ -1,5 +1,4 @@
 import type { ResponseError } from 'types'
-import { DEFAULT_QUERY_PARAMS } from './Reports.constants'
 
 export enum Presets {
   API = 'api',
@@ -11,7 +10,9 @@ export enum Presets {
 
 export type MetaQueryResponse = any & { error: ResponseError }
 
-export type BaseReportParams = typeof DEFAULT_QUERY_PARAMS & { sql?: string } & unknown
+export type BaseReportParams = { iso_timestamp_start: string; iso_timestamp_end: string } & {
+  sql?: string
+} & unknown
 export interface PresetConfig {
   title: string
   queries: BaseQueries<string>
