+++ apps/studio/components/interfaces/UnifiedLogs/UnifiedLogs.constants.tsx
@@ -15,23 +15,6 @@ import {
   SLIDER_DELIMITER,
   SORT_DELIMITER,
 } from 'components/ui/DataTable/DataTable.constants'
-import { ChartConfig } from 'ui'
-import { TooltipLabel } from './components/TooltipLabel'
-
-export const CHART_CONFIG = {
-  success: {
-    label: <TooltipLabel level="success" />,
-    color: 'hsl(var(--foreground-muted))',
-  },
-  warning: {
-    label: <TooltipLabel level="warning" />,
-    color: 'hsl(var(--warning-default))',
-  },
-  error: {
-    label: <TooltipLabel level="error" />,
-    color: 'hsl(var(--destructive-default))',
-  },
-} satisfies ChartConfig
 
 export const REGIONS = ['ams', 'fra', 'gru', 'hkg', 'iad', 'syd'] as const
 export const METHODS = ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'] as const
