+++ apps/studio/data/content/keys.ts
@@ -1,12 +1,9 @@
-import type { ContentType } from './content-query'
-import type { SqlSnippet } from './sql-snippets-query'
-
 export const contentKeys = {
   allContentLists: (projectRef: string | undefined) => ['projects', projectRef, 'content'] as const,
   infiniteList: (
     projectRef: string | undefined,
     options?: {
-      type: ContentType | undefined
+      type: string
       name: string | undefined
       limit?: number
       sort?: string
@@ -14,14 +11,14 @@ export const contentKeys = {
   ) => ['projects', projectRef, 'content-infinite', options].filter(Boolean),
   list: (
     projectRef: string | undefined,
-    options: { type?: ContentType; name?: string; limit?: number }
+    options: { type?: string; name?: string; limit?: number }
   ) => ['projects', projectRef, 'content', options] as const,
   sqlSnippets: (
     projectRef: string | undefined,
     options?: {
       sort?: 'inserted_at' | 'name'
       name?: string
-      visibility?: SqlSnippet['visibility']
+      visibility?: string
       favorite?: boolean
     }
   ) => ['projects', projectRef, 'content', 'sql', options].filter(Boolean),
@@ -41,7 +38,7 @@ export const contentKeys = {
     type?: string,
     options?: {
       cumulative?: boolean
-      visibility?: SqlSnippet['visibility']
+      visibility?: string
       favorite?: boolean
       name?: string
     }
