+++ apps/studio/data/table-rows/keys.ts
@@ -1,12 +1,5 @@
-import type { GetTableRowsArgs } from './table-rows-query'
-
-type TableRowKeyArgs = Omit<GetTableRowsArgs, 'table'> & { table?: { id?: number } }
-
 export const tableRowKeys = {
-  tableRows: (
-    projectRef?: string,
-    { table, roleImpersonationState, ...args }: TableRowKeyArgs = {}
-  ) =>
+  tableRows: (projectRef?: string, { table, roleImpersonationState, ...args }: any = {}) =>
     [
       'projects',
       projectRef,
@@ -15,7 +8,7 @@ export const tableRowKeys = {
       'rows',
       { roleImpersonation: roleImpersonationState?.role, ...args },
     ] as const,
-  tableRowsCount: (projectRef?: string, { table, ...args }: TableRowKeyArgs = {}) =>
+  tableRowsCount: (projectRef?: string, { table, ...args }: any = {}) =>
     ['projects', projectRef, 'table-rows', table?.id, 'count', args] as const,
   tableRowsAndCount: (projectRef?: string, tableId?: number) =>
     ['projects', projectRef, 'table-rows', tableId] as const,
